requirements.txt
pytz
python-dateutil==2.6.1
Scrapy==1.4.0
